requirements.txt
aiodns~=1.2.0
aiohttp~=3.5.4
gunicorn~=19.9.0
PyYAML~=3.13

git+https://github.com/shveenkov/aiotarantool.git
